+++ src/apps/ScopeChangeRequest/hooks/queries/useScopeChangeAccess.ts
@@ -1,13 +1,7 @@
 import { swap } from '@dbeining/react-atom';
 import { useQuery } from 'react-query';
-import { OptionRequestResult } from '../../api/ScopeChange/Access/optionsRequestChecker';
 import { scopeChangeAtom } from '../../Atoms/scopeChangeAtom';
 import { scopeChangeQueries } from '../../keys/queries';
-
-export interface ScopeChangeAccess extends OptionRequestResult {
-    canVoid: boolean;
-    canUnVoid: boolean;
-}
 
 export function useScopeChangeAccess(requestId: string): void {
     const { canUnvoidQuery, canVoidQuery, permissionsQuery } = scopeChangeQueries.permissionQueries;
