+++ src/apps/ScopeChangeRequest/types/query/optionsRequest.ts
@@ -1,0 +1,7 @@
+export type OptionRequestResult = {
+    canGet: boolean;
+    canPost: boolean;
+    canPut: boolean;
+    canPatch: boolean;
+    canDelete: boolean;
+};
