+++ src/apps/ScopeChangeRequest/types/query/scopeChangeAccess.ts
@@ -1,0 +1,6 @@
+import { OptionRequestResult } from './optionsRequest';
+
+export interface ScopeChangeAccess extends OptionRequestResult {
+    canVoid: boolean;
+    canUnVoid: boolean;
+}
